requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
bcrypt==4.2.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
diskcache==5.6.3
ecdsa==0.19.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
loguru==0.7.3
mutagen==1.47.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
py-mini-racer==0.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyExecJS==1.5.1
pyncm==1.7.1
python-jose==3.3.0
python-multipart==0.0.20
qrcode==8.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.34.0
